R39-DECOMP-VIEWER ③④: analyse + authoring out — the builders map is empty
app.ts 4,272 -> 3,953. Across the four slices: 5,160 -> 3,953, a 23% cut, and
`buildToolsPanel` no longer holds any section body at all.

The one thing that was NOT a clean accessor, stated rather than buried.

The analyse section does not merely READ mutable state, it WRITES it:
`dispose4d = populate4dPanel(...)`, then `dispose4d?.(); dispose4d = null` when
the panel closes. A getter cannot express a write, so the accessor rule is not
enough here.

Two wrong fixes were available. Moving the `let` into the new module changes its
LIFETIME — it lives in the viewer-app closure and must survive buildToolsPanel
running again on a persona change, which it does not do as a function-local.
Passing it by value is the exact stale-closure bug the accessor rule exists to
prevent. So ownership stays in app.ts and a mutable REF crosses the seam:
`fourD: { dispose: (() => void) | null }`.

That makes two lines of the moved body non-identical, and it is the ONLY such
exception in 1,180 moved lines. Verified rather than asserted: diffing each
moved body against `git show HEAD:app.ts` at its original range shows authoring
byte-identical, and analyse differing on exactly those two lines and nothing
else.

tsc again found what the census missed: waitForPublish, loadProjectModel and
openNodeCanvas in authoring; populate4dPanel and the LogisticsResource type in
analyse. Five more captures that would each have been a runtime failure.

toolsSplit.test.ts — the source-reading parity gate that caught ② — now lists
all four extracted files. Two mutations red: dropping an inventoried analyse
control, and forgetting to list analyseSection.ts in SECTION_SOURCES.

One correction to my own arithmetic: I pinned 3,952 from a script that counted
split-lines; test_file_sizes counts 3,953 and failed. The gate is the reader
that decides, so its number is the pin.

1307/1307 web green (116 files), tsc + eslint clean, test_file_sizes OK.

Next seam, for whoever takes it: buildPanels (152 lines) and the rail-panel
plumbing. buildToolsPanel is now scaffolding rather than content.
